+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `ids` to Styleguide components and HTML elements.
+
 ## [0.11.1] - 2019-10-16
 
 ### Fixed