+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.11.3] - 2019-10-23
+
 ### Added
 
 - "Remove" label to option zero in the quantity selector dropdown.